Update ErrorProne to 2.1.3 and fix failures
The fixes could have subtle side-effects, but I did take care when making them.

+++ benchmarks/src/main/java/io/grpc/benchmarks/qps/OpenLoopClient.java
@@ -137,7 +137,7 @@ public class OpenLoopClient {
       stub = BenchmarkServiceGrpc.newStub(checkNotNull(channel, "channel"));
       this.request = checkNotNull(request, "request");
       this.targetQps = targetQps;
-      numRpcs = targetQps * duration;
+      numRpcs = (long) targetQps * duration;
       rnd = new Random();
     }
